Supposed to be decimal datatype

+++ src/scripts/db/tables.sql
@@ -36,7 +36,7 @@ create table `accounts`
 	`id` int unsigned auto_increment,
 	`account_num` varchar(10) not null,
 	`user_id` int unsigned not null,
-	`balance` float(23, 2) default 0.00,
+	`balance` decimal(23, 2) default 0.00,
 
 	primary key (`id`),
 	foreign key (`user_id`) references `users` (`id`) on delete cascade on update restrict,
@@ -48,12 +48,12 @@ create table `transactions`
 	`id` int unsigned auto_increment,
 	`account_num` varchar(10) not null,
 	`time` datetime not null,
-	`amount` float(23, 2) not null,
+	`amount` decimal(23, 2) not null,
 	`transaction_type` enum("added", "sent", "received") not null,
 	`description` varchar(255),
 	`sender` varchar(10),
 	`recipient` varchar(64),
-	`new_balance` float(23,2) default 0.00
+	`new_balance` decimal(23, 2) default 0.00,
 
 	primary key (`id`),
 	foreign key (`account_num`) references `accounts` (`account_num`) on delete cascade on update restrict
